+++ core/chains/evm/gas/block_history_estimator_test.go
@@ -955,6 +955,11 @@ func TestBlockHistoryEstimator_Recalculate_NoEIP1559(t *testing.T) {
 		cfg.ChainTypeF = string(config.ChainGnosis)
 		bhe.Recalculate(cltest.Head(0))
 		require.Equal(t, assets.NewWeiI(80), gas.GetGasPrice(bhe))
+
+		// Same for xDai (deprecated)
+		cfg.ChainTypeF = string(config.ChainXDai)
+		bhe.Recalculate(cltest.Head(0))
+		require.Equal(t, assets.NewWeiI(80), gas.GetGasPrice(bhe))
 	})
 
 	t.Run("handles unreasonably large gas prices (larger than a 64 bit int can hold)", func(t *testing.T) {