+++ core/services/chainlink/config.go
@@ -12,6 +12,7 @@ import (
 	"github.com/smartcontractkit/chainlink-solana/pkg/solana"
 	stkcfg "github.com/smartcontractkit/chainlink-starknet/relayer/pkg/chainlink/config"
 
+	commoncfg "github.com/smartcontractkit/chainlink/v2/common/config"
 	evmcfg "github.com/smartcontractkit/chainlink/v2/core/chains/evm/config/toml"
 	"github.com/smartcontractkit/chainlink/v2/core/config/docs"
 	"github.com/smartcontractkit/chainlink/v2/core/config/env"
@@ -68,6 +69,16 @@ func (c *Config) valueWarnings() (err error) {
 			if c.Tracing.TLSCertPath != nil {
 				err = multierr.Append(err, config.ErrInvalid{Name: "Tracing.TLSCertPath", Value: *c.Tracing.TLSCertPath, Msg: "must be empty when Tracing.Mode is 'unencrypted'"})
 			}
+		}
+	}
+	// ChainType xdai is deprecated and has been renamed to gnosis
+	for _, evm := range c.EVM {
+		if evm.ChainType != nil && *evm.ChainType == string(commoncfg.ChainXDai) {
+			err = multierr.Append(err, config.ErrInvalid{
+				Name:  "EVM.ChainType",
+				Value: *evm.ChainType,
+				Msg:   "deprecated and will be removed in v2.12.0, use 'gnosis' instead",
+			})
 		}
 	}
 	return